+++ titan_client/api/entities_api.py
@@ -61,7 +61,7 @@ class EntitiesApi(object):
         :param sort: Sort results by relevance or by the object's native time in descending (latest) or ascending (earliest) order.
         :type sort: str
         :param offset: Skip leading number of records.
-        :type offset: float
+        :type offset: int
         :param count: Returns given number of records starting from `offset` position.
         :type count: int
         :param async_req: Whether to execute the request asynchronously.
@@ -107,7 +107,7 @@ class EntitiesApi(object):
         :param sort: Sort results by relevance or by the object's native time in descending (latest) or ascending (earliest) order.
         :type sort: str
         :param offset: Skip leading number of records.
-        :type offset: float
+        :type offset: int
         :param count: Returns given number of records starting from `offset` position.
         :type count: int
         :param async_req: Whether to execute the request asynchronously.
