+++ titan_client/api/forums_api.py
@@ -65,7 +65,7 @@ class ForumsApi(object):
         :param sort: Sort results by relevance or by the object's native time in descending (latest) or ascending (earliest) order.
         :type sort: str
         :param offset: Skip leading number of records.
-        :type offset: float
+        :type offset: int
         :param count: Returns given number of records starting from `offset` position.
         :type count: int
         :param async_req: Whether to execute the request asynchronously.
@@ -115,7 +115,7 @@ class ForumsApi(object):
         :param sort: Sort results by relevance or by the object's native time in descending (latest) or ascending (earliest) order.
         :type sort: str
         :param offset: Skip leading number of records.
-        :type offset: float
+        :type offset: int
         :param count: Returns given number of records starting from `offset` position.
         :type count: int
         :param async_req: Whether to execute the request asynchronously.
@@ -275,7 +275,7 @@ class ForumsApi(object):
         :param sort: Sort results by relevance or by the object's native time in descending (latest) or ascending (earliest) order.
         :type sort: str
         :param offset: Skip leading number of records.
-        :type offset: float
+        :type offset: int
         :param count: Returns given number of records starting from `offset` position.
         :type count: int
         :param async_req: Whether to execute the request asynchronously.
@@ -325,7 +325,7 @@ class ForumsApi(object):
         :param sort: Sort results by relevance or by the object's native time in descending (latest) or ascending (earliest) order.
         :type sort: str
         :param offset: Skip leading number of records.
-        :type offset: float
+        :type offset: int
         :param count: Returns given number of records starting from `offset` position.
         :type count: int
         :param async_req: Whether to execute the request asynchronously.
