+++ titan_client/api/girs_api.py
@@ -53,7 +53,7 @@ class GIRsApi(object):
         :param gir_description: Search GIRs description
         :type gir_description: str
         :param offset: Skip leading number of records.
-        :type offset: float
+        :type offset: int
         :param count: Returns given number of records starting from `offset` position.
         :type count: int
         :param async_req: Whether to execute the request asynchronously.
@@ -91,7 +91,7 @@ class GIRsApi(object):
         :param gir_description: Search GIRs description
         :type gir_description: str
         :param offset: Skip leading number of records.
-        :type offset: float
+        :type offset: int
         :param count: Returns given number of records starting from `offset` position.
         :type count: int
         :param async_req: Whether to execute the request asynchronously.
