+++ titan_client/api/messaging_services_api.py
@@ -67,7 +67,7 @@ class MessagingServicesApi(object):
         :param sort: Sort results by the object's native time in descending (latest) or ascending (earliest) order
         :type sort: str
         :param offset: Skip leading number of records.
-        :type offset: float
+        :type offset: int
         :param count: Returns given number of records starting from `offset` position.
         :type count: int
         :param async_req: Whether to execute the request asynchronously.
@@ -119,7 +119,7 @@ class MessagingServicesApi(object):
         :param sort: Sort results by the object's native time in descending (latest) or ascending (earliest) order
         :type sort: str
         :param offset: Skip leading number of records.
-        :type offset: float
+        :type offset: int
         :param count: Returns given number of records starting from `offset` position.
         :type count: int
         :param async_req: Whether to execute the request asynchronously.
